// Frontend/projectk-frontend/src/app/features/kurinModule/common/models/enums/plast-level.enum.ts
export enum PlastLevel {
    Entry = 'Entry',
    Uchasnyk = 'Uchasnyk',
    Rozviduvach = 'Rozviduvach',
    Skob = 'Skob',
    HetmanskiySkob = 'HetmanskiySkob',
    Starshoplastun = 'Starshoplastun',
    Senior = 'Senior',
    SeniorPratsi = 'SeniorPratsi',
    SeniorDovirja = 'SeniorDovirja',
    SeniorKerivnytstva = 'SeniorKerivnytstva'
}